Fix MCP test collection by moving tenancy fixtures to root plugin
Remove pytest_plugins from nested mcp/conftest.py and register shared
EvalHub tenant fixtures via tests/fixtures/evalhub_tenancy.py.
