+++ tests/ai_safety/evalhub/multitenancy/conftest.py
@@ -1,4 +1,3 @@
-import uuid
 from collections.abc import Generator
 from typing import Any
 
@@ -8,21 +7,16 @@ import structlog
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.deployment import Deployment
 from ocp_resources.namespace import Namespace
-from ocp_resources.role import Role
-from ocp_resources.role_binding import RoleBinding
 from ocp_resources.route import Route
 from ocp_resources.service import Service
-from ocp_resources.service_account import ServiceAccount
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
 from tests.ai_safety.evalhub.constants import (
     EVALHUB_HEALTH_PATH,
-    EVALHUB_TENANT_LABEL_KEY,
     EVALHUB_VLLM_EMULATOR_PORT,
 )
 from tests.ai_safety.evalhub.utils import tenant_rbac_ready
 from utilities.constants import Labels, Protocols, Timeout
-from utilities.infra import create_inference_token, create_ns
 
 LOGGER = structlog.get_logger(name=__name__)
 
@@ -61,45 +55,6 @@ def evalhub_mt_ready(
 
 
 # ---------------------------------------------------------------------------
-# Tenant namespaces
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture(scope="class")
-def tenant_a_namespace(
-    request: pytest.FixtureRequest,
-    admin_client: DynamicClient,
-) -> Generator[Namespace, Any, Any]:
-    """Tenant namespace where the test user HAS access."""
-    cls_name = request.cls.__name__.lower() if request.cls else "default"
-    suffix = uuid.uuid4().hex[:6]
-    name = f"test-evalhub-tenant-a-{cls_name}-{suffix}"
-    with create_ns(
-        admin_client=admin_client,
-        name=name,
-        labels={EVALHUB_TENANT_LABEL_KEY: "true"},
-    ) as ns:
-        yield ns
-
-
-@pytest.fixture(scope="class")
-def tenant_b_namespace(
-    request: pytest.FixtureRequest,
-    admin_client: DynamicClient,
-) -> Generator[Namespace, Any, Any]:
-    """Tenant namespace where the test user does NOT have access."""
-    cls_name = request.cls.__name__.lower() if request.cls else "default"
-    suffix = uuid.uuid4().hex[:6]
-    name = f"test-evalhub-tenant-b-{cls_name}-{suffix}"
-    with create_ns(
-        admin_client=admin_client,
-        name=name,
-        labels={EVALHUB_TENANT_LABEL_KEY: "true"},
-    ) as ns:
-        yield ns
-
-
-# ---------------------------------------------------------------------------
 # Wait for operator to provision tenant RBAC
 # ---------------------------------------------------------------------------
 
@@ -134,87 +89,6 @@ def tenant_a_rbac_ready(
         )
         LOGGER.error(msg)
         raise RuntimeError(msg) from err
-
-
-# ---------------------------------------------------------------------------
-# ServiceAccount and RBAC (only in tenant-a)
-# ---------------------------------------------------------------------------
-
-# Mirrors the user RBAC template from resources/evalhub-user-rbac-template.yaml.
-# evaluations/collections/providers are virtual SAR resources — not real CRDs.
-EVALHUB_USER_ROLE_RULES: list[dict[str, list[str]]] = [
-    {
-        "apiGroups": ["trustyai.opendatahub.io"],
-        "resources": ["evaluations", "collections", "providers", "status-events"],
-        "verbs": ["get", "list", "create", "update", "patch", "delete"],
-    },
-    {
-        "apiGroups": ["mlflow.kubeflow.org"],
-        "resources": ["experiments"],
-        "verbs": ["create", "get"],
-    },
-]
-
-
-@pytest.fixture(scope="class")
-def tenant_a_service_account(
-    admin_client: DynamicClient,
-    tenant_a_namespace: Namespace,
-) -> Generator[ServiceAccount, Any, Any]:
-    """ServiceAccount in tenant-a for multi-tenancy tests."""
-    with ServiceAccount(
-        client=admin_client,
-        name="evalhub-test-user",
-        namespace=tenant_a_namespace.name,
-        wait_for_resource=True,
-    ) as sa:
-        yield sa
-
-
-@pytest.fixture(scope="class")
-def tenant_a_evalhub_role(
-    admin_client: DynamicClient,
-    tenant_a_namespace: Namespace,
-) -> Generator[Role, Any, Any]:
-    """Role granting full EvalHub API access in tenant-a (virtual SAR resources)."""
-    with Role(
-        client=admin_client,
-        name="evalhub-test-user-access",
-        namespace=tenant_a_namespace.name,
-        rules=EVALHUB_USER_ROLE_RULES,
-        wait_for_resource=True,
-    ) as role:
-        yield role
-
-
-@pytest.fixture(scope="class")
-def tenant_a_evalhub_role_binding(
-    admin_client: DynamicClient,
-    tenant_a_namespace: Namespace,
-    tenant_a_service_account: ServiceAccount,
-    tenant_a_evalhub_role: Role,
-) -> Generator[RoleBinding, Any, Any]:
-    """RoleBinding granting the test SA EvalHub access in tenant-a only."""
-    with RoleBinding(
-        client=admin_client,
-        name="evalhub-test-user-binding",
-        namespace=tenant_a_namespace.name,
-        subjects_kind="ServiceAccount",
-        subjects_name=tenant_a_service_account.name,
-        role_ref_kind="Role",
-        role_ref_name=tenant_a_evalhub_role.name,
-        wait_for_resource=True,
-    ) as rb:
-        yield rb
-
-
-@pytest.fixture(scope="class")
-def tenant_a_token(
-    tenant_a_service_account: ServiceAccount,
-    tenant_a_evalhub_role_binding: RoleBinding,
-) -> str:
-    """Bearer token for the test SA (has access to tenant-a, not tenant-b)."""
-    return create_inference_token(model_service_account=tenant_a_service_account)
 
 
 # ---------------------------------------------------------------------------
